Added the "picture", "takepicture", "takePicture" command. Will take a picture and save to /sdcard/cellbot_%Y-%m-%d_%H-%M-%S.jpg Does not require user interaction since v21 of the ASE.
Added the whiteboard which is a log of sensor values and events. It is mainly for future development. 

ReaderThead parses input that contains a ':' and posts values onto the whiteboard.

Added the 'whiteboard' command which prints out the latest whiteboard values. 'whiteboardfull' will output the history of values too. 


If outputMethod is not in the config.ini file, prompt the use and save it.